Codes/Client_webapp/requirements.txt
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
filelock==3.3.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Excel==0.0.7
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-response==0.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
huggingface-hub==0.0.17
idna==3.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==7.0.3
jupyter-core==4.8.1
lml==0.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mysql==0.0.3
mysqlclient==2.0.3
nest-asyncio==1.5.1
nltk==3.6.3
numpy==1.21.2
packaging==21.0
pampy==0.3.0
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
pycparser==2.20
pyexcel==0.6.7
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.9.30
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
SQLAlchemy==1.4.23
texttable==1.6.4
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.11.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
